Overall messaging cleanup. Factored messaging related code out of s3.py and sync.py into messaging.py. Added a Message class to replace the message tuple.
