+++ calcloud/s3.py
@@ -315,149 +315,3 @@ def delete_object(s3_filepath, client=None):
     client, bucket_name, object_name = _s3_setup(client, s3_filepath)
     return client.delete_object(Bucket=bucket_name, Key=object_name)
 
-# -------------------------------------------------------------
-
-
-class S3Messenger:
-    """S3Messenger handles sending and receiving messages via files in S3.  Messages
-    can contain arbitrary text and when sent are uploaded as S3 objects stored
-    at a path structured like this:
-
-       `s3_path`/`message_type`/`message_name`
-
-    S3 `message_type`s correspond roughly to message queues to which messages are
-    sent.   To ease implementation,  the names of messages are preserved and available.
-
-    `s3_path`        batch output path.  Extended by "messages" and "data."
-    `messaging_type` the kind of message, e.g. dataset-processed.
-    `message_name`   the S3 object where message is stored, e.g. an ipppssoot.
-    """
-    def __init__(self, s3_path):
-        self.s3_path = s3_path
-        self.client = boto3.client("s3")
-        self.tmp_dir = tempfile.gettempdir()
-        if not os.path.isdir(self.tmp_dir):
-            os.mkdir(self.tmp_dir)
-
-    def message_path(self, message_type, message_name):
-        """Given strings `message_type` and `message_name` return the
-        S3 path for that message.
-        """
-        return "/".join([self.s3_path, "messages", message_type, message_name])
-
-    def data_path(self, name):
-        """Given string  `name`,  return the corresponding S3 path within
-        the messenger's data branch.
-        """
-        return "/".join([self.s3_path, "data", name])
-
-    def message_name(self, s3_msg_path):
-        """Given `s3_msg_path` string,  return the name of the message."""
-        return s3_msg_path.split("/")[-1]
-
-    def download_directory(self, local_path, s3_path, max_files=1000):
-        """Given S3 directory at `s3_path`,  download all files to `local_path`."""
-        log.info(f"Downloading '{s3_path}' to '{local_path}'.")
-        return download_directory(
-            local_path, s3_path, client=self.client, max_objects=max_files)
-
-    def list_messages(self, message_type, max_messages=1):
-        """Given string `message_type`,  return the S3 paths of up to
-        `max_messages` of that  type.
-        """
-        s3_prefix = self.s3_path + "/messages/" + message_type
-        log.verbose(f"Listing messages at '{s3_prefix}'.")
-        return list_directory(
-            s3_prefix, client=self.client, max_objects=max_messages)
-
-    def list_names(self, message_type, max_messages=1):
-        """Given string `message_type`,  return the message names of
-        up to `max_messages` of that type.
-        """
-        paths = self.list_messages(message_type, max_messages)
-        return [self.message_name(path) for path in paths]
-
-    def receive(self, message_type, encoding='utf-8'):
-        """Given string `message_type`, fetch one message of that
-        type and delete the S3 copy.   Return the name of the
-        message and contents loaded from the S3 file.
-
-        Returns (message_name, message_content) or ()
-        """
-        message_paths = self.list_messages(message_type)
-        if message_paths:
-            message = self.load(message_paths[0], encoding)
-            delete_object(message_paths[0], self.client)
-            return message
-        return ()
-
-    def load(self, message_path, encoding='utf-8'):
-        """Given string S3 `message_path`,  get the object contents
-        from that message path with `encoding`.
-
-        return (message_name, message_contents)
-        """
-        contents = get_object(
-            message_path, client=self.client, encoding=encoding)
-        return (self.message_name(message_path), contents)
-
-    def send(self, message_type, message):
-        """Given string `message_type` and a (name, contents) tuple
-        `message`,  upload the contents to that  appropriate S3 path
-        for that message type and name.
-        """
-        message_name, message_contents = message
-        tmp_name = os.path.join(self.tmp_dir, message_name)
-        with open(tmp_name, "w+") as _tmp_file:
-            _tmp_file.write(message_contents)
-        s3_name = self.message_path(message_type, message_name)
-        upload_filepath(tmp_name, s3_name, client=self.client)
-
-    def receive_and_resend(self, old_type, new_type):
-        """Given S3 message type string  `old_type` and `new_type`,
-        receive one message from `old_type`,  send it to `new_type`,  and
-        return it.   This corresponds to moving the message from one
-        queue to the next in a linear series while reading the value.
-
-        Returns  (message_name, message_contents)
-        """
-        message = self.receive(old_type)
-        if message:
-            self.send(new_type, message)
-        return message
-
-    def delete(self, message_type, message_name):
-        """Given  string S3 `message_type` and `message_name`, delete
-        the S3 object at the corresponding S3 message path.
-
-        This is essentially the same thing as removing a message from
-        an AWS SQS queue after receiving it where `message_name` has
-        an intelligble value.
-        """
-        path = self.message_path(message_type, message_name)
-        delete_object(path, self.client)
-
-    def pass_message(self, old_type, new_type, message):
-        """Given S3 message type string `old_type`, `new_type`, and
-        (name, contents) tuple `message`,  send the contents
-        of `message` to `new_type` and delete the object at S3
-        `old_type` with the message's `name`.
-
-        Assuming `contents` match the message with `name` of `old_type`,
-        this corresponds to moving the message from queue `old_type` to
-        queue `new_type`.
-        """
-        self.send(new_type, message)
-        self.delete(old_type, message[0])
-
-    def reset(self, to_type, from_types):
-        """Given message type `to_type`,  find all S3 messages for the types
-        in list `from_types` and move them to `to_type`.   This can be used
-        to reset batch outputs to their original un-synced state following
-        failures or during development.
-        """
-        for from_type in from_types:
-            for path in self.list_messages(from_type, max_messages=2**30):
-                self.pass_message(from_type, to_type, self.load(path))
-
-
